% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{farm.adjust}
\alias{farm.adjust}
\title{Adjusting a data matrix for underlying factors}
\usage{
farm.adjust(X, Y, factors = NULL, K.factors = NULL, robust.cov = FALSE)
}
\arguments{
\item{X}{a n x p data matrix with each row being a sample.}

\item{Y}{a n x 1 outcome vector.}

\item{factors}{an \emph{optional} factor matrix with each column being a factor for \code{X}. Same number of rows as \code{X}.}

\item{K.factors}{a \emph{optional} number of factors to be estimated. Otherwise estimated internally.}

\item{robust.cov}{an \emph{optional} boolean, specifying whether or not to use a robust covariance matrix for factor adjustment. Default is FALSE.}
}
\value{
A list with the following items
\item{residual}{the data after being adjusted for underlying factors}
\item{loadings}{estimated factor loadings}
\item{factors}{estimated factors}
\item{nfactors}{if needed, the number of estimated factors}
}
\description{
Given a matrix of covariates, this function estimates the underlying factors and computes data residuals after regressing out those factors.
}
\details{
Using \code{robust.cov = TRUE} uses the Huber's loss to estimate the covariance matrix. For details of covariance estimation method see Zhou et al.(2017).

Number of rows and columns of the data matrix must be at least 4 in order to be able to calculate latent factors.
}
\references{
Zhou W.-X., Fan J., Ke Y., Sun Q., "FARM-Test: Factor-adjusted robust multiple testing with false discovery control." \url{https://goo.gl/68SJpd}
}
