% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{farm.select}
\alias{farm.select}
\title{Factor-adjusted robust model selection}
\usage{
farm.select(Y, X, loss = c("lasso", "scad", "mcp"), verbose = TRUE, ...)
}
\arguments{
\item{Y}{a n x 1 outcome vector.}

\item{X}{a n x p covariate matrix with each row being a sample. Must have same number of rows as \code{Y}.}

\item{loss}{an \emph{optional} character string specifying the loss function to be minimized. Must be one of "lasso" (default), "scad" or "mcp". You can just specify the initial letter.}

\item{verbose}{an \emph{optional} logical determining whether or not to print output to the console. Default is TRUE.}

\item{\dots}{Arguments passed to the \code{\link{farm.adjust}} function.}
}
\value{
A list with the following items
\item{beta.chosen}{the indices of the covariates chosen in the model}
}
\description{
Given a covariate matrix and output vector, this function first adjusts the covariates for underlying factors and then performs model selection.
}
\details{
Number of rows and columns of the data matrix must be at least 4 in order to be able to calculate latent factors.

For details about factor adjustment, see !!! Kaizheng's paper !!
}
\examples{

}
